+++ lib/coherence/plugs/require_login.ex
@@ -3,7 +3,7 @@ defmodule Coherence.RequireLogin do
   Plug to protect controllers that require login.
   """
 
-  import Coherence.ControllerHelpers, only: [logged_out_url: 1]
+  import Coherence.Controller, only: [logged_out_url: 1]
   import Plug.Conn
   import Phoenix.Controller, only: [put_flash: 3, redirect: 2]