+++ lib/coherence/schemas/rememberable.ex
@@ -68,10 +68,10 @@ defmodule Coherence.Rememberable do
       def created_at, do: Timex.now
 
       def gen_token do
-        Coherence.ControllerHelpers.random_string 24
+        Coherence.Controller.random_string 24
       end
       def gen_series do
-        Coherence.ControllerHelpers.random_string 10
+        Coherence.Controller.random_string 10
       end
 
       defoverridable [